requirements.txt
minio
fastapi
uvicorn
docxtpl==0.8.1
python-docx==0.8.10
